Subida del cuarto ejercicio

+++ Interview4/ItemStack.h
@@ -8,19 +8,19 @@ class ItemStack
 {
   public:
     ItemStack() = delete;
-    ItemStack(const std::string& title, const float& cost, const unsigned int& count);
+    ItemStack(const std::string& title, const double& cost, const unsigned int& count);
     std::string title() const;
-    float cost() const;
+    double cost() const;
     unsigned int count() const;
 
     void setTitle(const std::string& title);
-    void setCost(const float& cost);
+    void setCost(const double& cost);
     void setCount(const unsigned int& count);
     void retreive();
     friend std::ostream& operator<< (std::ostream& stream, const ItemStack& itemStack);
   private:
     std::string m_title;
-    float m_cost; 
+    double m_cost; 
     unsigned int m_count;
 };
